made refactoring, added color accent to the active page nav link
