+++ src/components/shared/header/Header.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import HeaderDesk from "./HeaderDesk";
-import HeaderMob from "./HeaderMob";
+import HeaderMob from "./socialLinks/headerMob/HeaderMob";
 
 export default function Header() {
   return (
