+++ src/components/shared/header/navMenu/NavMenu.tsx
@@ -1,5 +1,6 @@
 "use client";
 import { useTranslations } from "next-intl";
+import { usePathname } from "@/i18n/routing";
 import React, { Dispatch, SetStateAction } from "react";
 import MenuLink from "./MenuLink";
 
@@ -9,6 +10,8 @@ interface NavMenuProps {
 
 export default function NavMenu({ setIsHeaderMenuOpened }: NavMenuProps) {
   const t = useTranslations();
+
+  const currentPath = usePathname().slice(1);
 
   const menuList = [
     { title: t("header.navMenu.home"), path: "" },
@@ -26,6 +29,7 @@ export default function NavMenu({ setIsHeaderMenuOpened }: NavMenuProps) {
             key={idx}
             menuItem={menuItem}
             setIsHeaderMenuOpened={setIsHeaderMenuOpened}
+            className={`${currentPath === menuItem.path ? "text-yellow" : ""}`}
           />
         ))}
       </ul>
